+++ copytrading/futures_follower_config.go
@@ -23,9 +23,9 @@ import (
 	"context"
 	"net/url"
 
+	copytypes "github.com/tonymontanov/go-bitget/v2/copytrading/types"
 	"github.com/tonymontanov/go-bitget/v2/internal/bgcommon"
 	"github.com/tonymontanov/go-bitget/v2/internal/rest"
-	copytypes "github.com/tonymontanov/go-bitget/v2/copytrading/types"
 )
 
 // ---------------------------------------------------------------------